requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
colorama==0.4.6
datasets==2.12.0
dill==0.3.6
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
lit==16.0.2
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.6
packaging==23.1
pandas==2.0.1
pick==2.2.0
protobuf==4.22.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
simple-term-menu==1.6.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
torch==2.0.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.3
xxhash==3.2.0
yarl==1.9.2
